Setup P as a hotkey to launch a game
There is a new launch_game classmethod that can be implemented for
RenderEngines that want to take advantage of this. This hotkey/operator
only works while the RenderEngine is selected and will fallback to its
default behavior (run BGE) if another RenderEngine is selected.

Any exception raised by launch_game() will be reported using Blender's
report system (i.e., it will show up in the GUI).